requirements.txt
aiogram
akinator.py